# server/server/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package server

type BearerAuth struct {
	Token string
	Roles []string
}

// GetToken returns the value of Token.
func (s *BearerAuth) GetToken() string {
	return s.Token
}

// GetRoles returns the value of Roles.
func (s *BearerAuth) GetRoles() []string {
	return s.Roles
}

// SetToken sets the value of Token.
func (s *BearerAuth) SetToken(val string) {
	s.Token = val
}

// SetRoles sets the value of Roles.
func (s *BearerAuth) SetRoles(val []string) {
	s.Roles = val
}

type CreateRbacResourceBadRequest ErrorMessage

func (*CreateRbacResourceBadRequest) createRbacResourceRes() {}

type CreateRbacResourceInternalServerError ErrorMessage

func (*CreateRbacResourceInternalServerError) createRbacResourceRes() {}

type CreateRbacResourceUnauthorized ErrorMessage

func (*CreateRbacResourceUnauthorized) createRbacResourceRes() {}

// Ref: #/ErrorMessage
type ErrorMessage struct {
	Message string `json:"message"`
}

// GetMessage returns the value of Message.
func (s *ErrorMessage) GetMessage() string {
	return s.Message
}

// SetMessage sets the value of Message.
func (s *ErrorMessage) SetMessage(val string) {
	s.Message = val
}

func (*ErrorMessage) getUnprotectedResourceRes() {}

type GetMeInternalServerError ErrorMessage

func (*GetMeInternalServerError) getMeRes() {}

type GetMeUnauthorized ErrorMessage

func (*GetMeUnauthorized) getMeRes() {}

type GetProtectedResourceInternalServerError ErrorMessage

func (*GetProtectedResourceInternalServerError) getProtectedResourceRes() {}

type GetProtectedResourceUnauthorized ErrorMessage

func (*GetProtectedResourceUnauthorized) getProtectedResourceRes() {}

type GetRbacResourceInternalServerError ErrorMessage

func (*GetRbacResourceInternalServerError) getRbacResourceRes() {}

type GetRbacResourceNotFound ErrorMessage

func (*GetRbacResourceNotFound) getRbacResourceRes() {}

type GetRbacResourceUnauthorized ErrorMessage

func (*GetRbacResourceUnauthorized) getRbacResourceRes() {}

// A protected resource that any authenticated user can access.
// Ref: #/ProtectedResource
type ProtectedResource struct {
	// The message of the resource.
	Message string `json:"message"`
	// The secret message of the resource.
	SomethingSecret string `json:"something_secret"`
}

// GetMessage returns the value of Message.
func (s *ProtectedResource) GetMessage() string {
	return s.Message
}

// GetSomethingSecret returns the value of SomethingSecret.
func (s *ProtectedResource) GetSomethingSecret() string {
	return s.SomethingSecret
}

// SetMessage sets the value of Message.
func (s *ProtectedResource) SetMessage(val string) {
	s.Message = val
}

// SetSomethingSecret sets the value of SomethingSecret.
func (s *ProtectedResource) SetSomethingSecret(val string) {
	s.SomethingSecret = val
}

func (*ProtectedResource) getProtectedResourceRes() {}

type Rbac struct {
	Token string
	Roles []string
}

// GetToken returns the value of Token.
func (s *Rbac) GetToken() string {
	return s.Token
}

// GetRoles returns the value of Roles.
func (s *Rbac) GetRoles() []string {
	return s.Roles
}

// SetToken sets the value of Token.
func (s *Rbac) SetToken(val string) {
	s.Token = val
}

// SetRoles sets the value of Roles.
func (s *Rbac) SetRoles(val []string) {
	s.Roles = val
}

// A protected resource that requires rbac.
// Ref: #/RbacResource
type RbacResource struct {
	// The message of the resource.
	Message string `json:"message"`
	// The secret message of the resource.
	SomethingSecret string `json:"something_secret"`
	// The very secret message of the resource.
	SomethingVerySecret string `json:"something_very_secret"`
}

// GetMessage returns the value of Message.
func (s *RbacResource) GetMessage() string {
	return s.Message
}

// GetSomethingSecret returns the value of SomethingSecret.
func (s *RbacResource) GetSomethingSecret() string {
	return s.SomethingSecret
}

// GetSomethingVerySecret returns the value of SomethingVerySecret.
func (s *RbacResource) GetSomethingVerySecret() string {
	return s.SomethingVerySecret
}

// SetMessage sets the value of Message.
func (s *RbacResource) SetMessage(val string) {
	s.Message = val
}

// SetSomethingSecret sets the value of SomethingSecret.
func (s *RbacResource) SetSomethingSecret(val string) {
	s.SomethingSecret = val
}

// SetSomethingVerySecret sets the value of SomethingVerySecret.
func (s *RbacResource) SetSomethingVerySecret(val string) {
	s.SomethingVerySecret = val
}

func (*RbacResource) createRbacResourceRes() {}
func (*RbacResource) getRbacResourceRes()    {}

// An unprotected resource.
// Ref: #/UnprotectedResource
type UnprotectedResource struct {
	// The message of the resource.
	Message string `json:"message"`
}

// GetMessage returns the value of Message.
func (s *UnprotectedResource) GetMessage() string {
	return s.Message
}

// SetMessage sets the value of Message.
func (s *UnprotectedResource) SetMessage(val string) {
	s.Message = val
}

func (*UnprotectedResource) getUnprotectedResourceRes() {}

// A user.
// Ref: #/User
type User struct {
	// The user's id.
	UserID string `json:"user_id"`
	// The user's first name.
	FirstName string `json:"first_name"`
	// The user's last name.
	LastName string `json:"last_name"`
	// The user's email.
	Email string `json:"email"`
}

// GetUserID returns the value of UserID.
func (s *User) GetUserID() string {
	return s.UserID
}

// GetFirstName returns the value of FirstName.
func (s *User) GetFirstName() string {
	return s.FirstName
}

// GetLastName returns the value of LastName.
func (s *User) GetLastName() string {
	return s.LastName
}

// GetEmail returns the value of Email.
func (s *User) GetEmail() string {
	return s.Email
}

// SetUserID sets the value of UserID.
func (s *User) SetUserID(val string) {
	s.UserID = val
}

// SetFirstName sets the value of FirstName.
func (s *User) SetFirstName(val string) {
	s.FirstName = val
}

// SetLastName sets the value of LastName.
func (s *User) SetLastName(val string) {
	s.LastName = val
}

// SetEmail sets the value of Email.
func (s *User) SetEmail(val string) {
	s.Email = val
}

func (*User) getMeRes() {}